ResourceIdentifiable now derivable, needs id field type inference
Basic implementation of derivable ResourceIdentifiable, only works
for id fields that has type String, or have the method
`.to_string()`. Should support built-in types (&str, float, etc.)
which is probably best implemented with a macro that takes two
arguments, first is id identifier, second is the type, and the
macro does a simple pattern match on the second argument to
determine what code should be emitted.

Macro hygiene is not a consideration right now, although it is
noted in the TODO.md document, keeping track of what needs to be
done.