Redefine ResourceIdentifiable to accept Id types that can be strings, applied rustfmt
The ResourceIdentifiable `get_id()` only supported returning a reference
to an Id field of type String (thereby becomming &str). This design
allows returning any type that can be turned into a string with
ToString.

By making the return type of `get_id()` a reference, we can simply read
the value, and no copying needs to occur. With specialization we can
even optimise the case where the Id type is a String, and simply read
the value instead of calling `to_string()`.

Derive code is broken though...