Proper serializing of ResourceIdentifiable through newtypes, remove JsonPrimaryDatatObject struct, using trait based approach instead, cleanup
Newtypes are a zero-cost abstraction to give some type a new type. By
implementing serialize on these newtypes, we can customize seralization
of data structs that have ResourceIdentifiable implemented.

Specialization is used to differentiate between serializing a single
data ojbect, and a vector of data objects. Specialization when Parsing
ResourceIdentifiable when IdType is a string is also used, so we only
read the id when it is already a string.

JsonPrimaryDataObject struct has been redefined to a newtype, that is a
private type, so the user simply provides the data without having to
construct anything.

JsonApiResponse has therefore been redefined to be a newtype of a result
of user provided data or user provided error. This is cleaner, and a
step closer to the trait based approach we want.