Setup rough derive macro error framework
The idea is to allow the expand_proc_macro wrapper function to expand
any deriviation with any Error type specialized for that particular
parser.

It is not decided whether specialized Error types will be used for each
derive macro, or a general Error type will be defined. The ability to
choose freely is now there.