Add specialization for serializing resource objects id fields of of all types that can be cast as &str
When serializing resource objects, the spec specifies that an id field
must exist, and must be a string. Our trait definition for resource
objects defines the id as a reference to a type that implements
`ToString`. When serializing resource objects, we get a reference to the
id and call `to_string()` on it.

Previously we had a specialization for the case where the id type was a
`String`, to not call `to_string()`, but simply serialize as is.
However, many other types, like `&str` and `Box<String>` etc. could also
use the same optimization.

When serializing id fields, we don't always need to call to_string(), if
the id field is already a type of string. If we can cast the type to
`&str` we can serialize the type as is, no need to convert or copy.

Changing the specialization to a self implemented trait for types that
can be cast to `&str`, we can specialize for that trait instead, giving
us the desired behaviour.

Maybe an already existing trait (`Into<_>`?) can be used to specialize
for cases where we can cast to `&str`, but this works as well for now.