Write tests for JsonApiResponse, cleanup, rename to ResourceIdentifierObject, implement newtype to serialize ResourceIdentifiable as `ResourceIdentifierObject`
Renaming `ResourceIdentifier` to `ResourceIdentifierObject`, which is
the correct name from the specification for the type of object we construct.

If we want to create resource linkage from some data object with
ResourceIdentifierObject implemented, we can simply serialize create a
custom serialize implementation with a newtype wrapper. This wrapper has
been called `ToResourceIdentifierObject`

Now testing that serializing of JsonApiResponse return correct top-level
document