Create error module, build representation for errors in JSON:API, macro for initializing variable amount of fields
Errors in JSON:API is an array of error objects, which are objects with
a set of optional fields. The endpoint must also return some HTTP error
code.

A struct holds the error code and the array of error objects. The error
objects are a struct with fields of Option<String> corresponding to the
fields an error may contain per the specification.

A macro `json_api_error` makes it easy to initialize this error obejct
without defining all the unused fields as None.