Split ResourceIdentifiable into ResourceType trait, implement basic deserialization for input requests
In order to be able to distinguish between resource objects that are being created, and does not have an ID yet,
we split ResourceIdentifiable into ResourceType so the type of a resource object can be known without the ID.

Basic implementation for deserializing a resource object as an input data request has been implemented.
Still lots of questions to be answered in terms of best practices when we want to use the same type,
potentially with an ID field, everywhere. How do we support client ids? How about PATCH updates?