ResourceType and ResourceIdentifiable move to own module, generalized responses, added a create response enum for each type of acceptable response
JsonApiResponse was renamed to JsonApiDataResponse. This should better
reflect its use, since we now have introduced JsonApiCreateResponse.

ResourceType and ResourceIdentifiable have now been moved to the module
resource. This should provide clarity, and response was getting crowded.

JsonApiCreateResponse has been added as a responder, it is an enum with
the different response types typically needed for create statements.

A generalized json api response was made in the core module
general_response, and JsonApiCreateResource now shares code with
JsonApiDataResponse.