Make ResourceType->get_type a static method
The resource type of an object is static, and so the method get_type has
been changed to be static. This created some problems for
ResourceIdentifierObject, because it has no type. Its definition has now
been changed, so it now has a generic of the data it represents, but
without any overhead (PhantomData).

This however created problems for RelationObject, because it needs to be
non-generic, so we can create a data structure of different
RelationObjects, to support having relationships to many different
types.

We now do a mapping of ResourceIdentifierObject to an identical
structure that does not implement ResourceType and ResourceIdentifiable.
This may seem a roundabout way, why not just remove these
implementations for ResourceIdentifierObject? This is because the
specifiation states that it is allowed to use resource identifier
objects as primary data, so we will probably move it to the public API.